Move the detail search_n implementation in detail
